# src/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Это API для управления пользователями, товарами и поставщиками.",
        "title": "API документация",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/image/create": {
            "post": {
                "description": "Создаёт новое изображение и привязывает его к продукту",
                "tags": [
                    "images"
                ],
                "summary": "Создать изображение",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/response.UploadUpdateImage"
                            }
                        }
                    },
                    "description": "Данные изображения",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/image/delete/{id}": {
            "delete": {
                "description": "Удаляет изображение по ID",
                "tags": [
                    "images"
                ],
                "summary": "Удалить изображение",
                "parameters": [
                    {
                        "description": "UUID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/image/product/{id}": {
            "get": {
                "description": "Возвращает изображение по UUID продукта",
                "tags": [
                    "images"
                ],
                "summary": "Получить изображение по ID продукта",
                "parameters": [
                    {
                        "description": "UUID продукта",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/image/updateImage": {
            "put": {
                "description": "Обновляет данные изображения по его ID",
                "tags": [
                    "images"
                ],
                "summary": "Обновить изображение",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/response.UploadUpdateImage"
                            }
                        }
                    },
                    "description": "Обновленные данные изображения",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/image/{id}": {
            "get": {
                "description": "Возвращает изображение по UUID",
                "tags": [
                    "images"
                ],
                "summary": "Получить изображение по его ID",
                "parameters": [
                    {
                        "description": "UUID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:5000/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "response.UploadUpdateImage": {
                "type": "object",
                "properties": {
                    "image": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "product_id": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}